requirements.txt
numpy
opencv-python
keras
Pillow
